feat(ai, sales): Integrate AI for sales analysis & refactor core models
This commit introduces the initial AI integration for sales analysis and performs a significant refactoring of core data models and database schema.

Key changes include:

*   **AI Integration (RAG)**:
    *   Added `AIService` to generate and store embeddings for sales data.
    *   Implemented a new API endpoint (`/api/v1/assistant/chat`) for natural language queries against sales data using RAG.
    *   Ensured AI processing runs as a background task with an independent DB session for robustness.
*   **Sales Module Enhancements**:
    *   Refactored `SaleService` to trigger AI embedding creation upon new sales.
    *   Added comprehensive unit tests for `SaleService.create_sale` covering success, no-items, and product-not-found scenarios.
*   **Data Model & Schema Refinements**:
    *   Standardized all monetary fields (`price`, `amount`, `balance`) to `Decimal` across `Sale`, `Purchase`, `Product`, `Shift`, `AccountsPayable`, `AccountsReceivable`, and `CashAccount` models/schemas to ensure financial precision.
    *   Corrected and expanded `Sale` model and `sales` table to include `user_id`, `store_id`, `pos_terminal_id`, `total_tax_amount`, `discount_amount`, `status`, `payment_status`, `shift_id`, and audit fields for completeness and consistency.
    *   Introduced `SaleItem` model and `sale_items` table to properly detail sale contents.
    *   Refactored `CashAccount` to represent a static account and introduced `CashTransaction` model/table for dynamic money movements within shifts.
    *   Updated `Shift` model and schemas to correctly use `starting_cash` and `ending_cash` with `Decimal` types.
    *   Ensured all necessary `Decimal` imports are present in relevant model files.
*   **Database Schema Updates (`V1__initial_schema.sql`)**:
    *   Updated `sales` and `cash_accounts` table definitions.
    *   Added `sale_items` and `cash_transactions` tables.
    *   Defined `cash_transaction_type` enum.
*   **Infrastructure Improvements**:
    *   Standardized `pos-db` container name in `docker-compose.yml`.
    *   Adjusted `OLLAMA_BASE_URL` in `docker-compose.yml` for better host communication.
    *   Fixed redundant prefix in `main.py` for `assistant_controller`.

These changes significantly enhance the API's functionality, data integrity, and maintainability, laying a solid foundation for future intelligent features.

+++ backend/app/models/accounts_payable.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from core.db import Base
@@ -35,8 +36,8 @@ class AccountsPayable(Base):
     supplier_id: Mapped[uuid.UUID] = mapped_column(
         PG_UUID(as_uuid=True), ForeignKey("suppliers.id"), nullable=False
     )
-    original_amount: Mapped[float] = mapped_column(Numeric(18, 2), nullable=False)
-    outstanding_amount: Mapped[float] = mapped_column(Numeric(18, 2), nullable=False)
+    original_amount: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
+    outstanding_amount: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
     due_date: Mapped[date | None] = mapped_column(DATE)
     status: Mapped[str] = mapped_column(String, nullable=False, default="OPEN")
