+++ backend/app/models/product.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import uuid
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from core.db import Base
@@ -22,7 +23,7 @@ class Product(Base):
     )
     name: Mapped[str] = mapped_column(String(255), nullable=False, index=True)
     description: Mapped[str | None] = mapped_column(String)
-    price: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    price: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     sku: Mapped[str | None] = mapped_column(String(100), unique=True, index=True)
 
     # Foreign Key to Supplier
