+++ backend/app/schemas/accounts_payable.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -11,8 +12,8 @@ class AccountsPayableBase(BaseModel):
 
     purchase_id: uuid.UUID
     supplier_id: uuid.UUID
-    original_amount: float
-    outstanding_amount: float
+    original_amount: Decimal
+    outstanding_amount: Decimal
     due_date: date | None = None
     status: ARAPStatus = ARAPStatus.OPEN
 
@@ -24,7 +25,7 @@ class AccountsPayableCreate(AccountsPayableBase):
 class AccountsPayableUpdate(BaseModel):
     """Esquema para actualizar una cuenta por pagar. Todos los campos son opcionales."""
 
-    outstanding_amount: float | None = None
+    outstanding_amount: Decimal | None = None
     due_date: date | None = None
     status: ARAPStatus | None = None
